Move sandbox resource tests out into their own files.
